+++ src/auto/info_bar.rs
@@ -472,6 +472,10 @@ pub trait InfoBarExt: 'static {
 
     //fn add_buttons(&self, first_button_text: &str, : /*Unknown conversion*//*Unimplemented*/Fundamental: VarArgs);
 
+    fn get_action_area(&self) -> Option<Box>;
+
+    fn get_content_area(&self) -> Box;
+
     fn get_message_type(&self) -> MessageType;
 
     #[cfg(any(feature = "v3_22_29", feature = "dox"))]
@@ -534,6 +538,22 @@ impl<O: IsA<InfoBar>> InfoBarExt for O {
     //fn add_buttons(&self, first_button_text: &str, : /*Unknown conversion*//*Unimplemented*/Fundamental: VarArgs) {
     //    unsafe { TODO: call gtk_sys:gtk_info_bar_add_buttons() }
     //}
+
+    fn get_action_area(&self) -> Option<Box> {
+        unsafe {
+            from_glib_none(gtk_sys::gtk_info_bar_get_action_area(
+                self.as_ref().to_glib_none().0,
+            ))
+        }
+    }
+
+    fn get_content_area(&self) -> Box {
+        unsafe {
+            from_glib_none(gtk_sys::gtk_info_bar_get_content_area(
+                self.as_ref().to_glib_none().0,
+            ))
+        }
+    }
 
     fn get_message_type(&self) -> MessageType {
         unsafe {
