+++ src/auto/versions.txt
@@ -1,2 +1,2 @@
 Generated by gir (https://github.com/gtk-rs/gir @ e75bde2)
-from gir-files (https://github.com/gtk-rs/gir-files @ c31e734)
+from gir-files (https://github.com/gtk-rs/gir-files @ 3043b14)
