+++ src/lib.rs
@@ -225,7 +225,6 @@ mod fixed;
 #[cfg(any(feature = "v3_18", feature = "dox"))]
 mod flow_box;
 mod im_context_simple;
-mod info_bar;
 mod invisible;
 #[cfg(any(feature = "v3_16", feature = "dox"))]
 mod list_box;
@@ -255,7 +254,6 @@ mod target_list;
 mod text_buffer;
 mod text_iter;
 mod tree_model_filter;
-mod tree_model_sort;
 mod tree_path;
 mod tree_row_reference;
 mod tree_sortable;
